Debugueo I
Detección de cerrado de conexión
client: no getteaba las colas, usaba id's incorrectos
broker: Terminado correcto de procesos hijos
worker: if sin comparación para recibir de cola
socket.c: Reutilizado de direcciones. Resuelvo warnings
Emprolijado general, loggeo en color, más mensajes de error

+++ common/ipc/socket.h
@@ -1,8 +1,8 @@
 #ifndef SOCKET_H
 #define SOCKET_H
 
-int create_client_socket(const char* server_ip, int server_port);
-int create_server_socket(int client_port);
+int create_client_socket(const char* server_ip, uint16_t server_port);
+int create_server_socket(uint16_t client_port);
 int accept_client(int server_socket);
 
 #endif //SOCKET_H
